+++ docs/app/(examples)/pattern/filter/utils.tsx
@@ -1,7 +1,27 @@
-import { venueTypes, venues } from '@/lib/data/venues';
+import { venueTraits, venueTypes, venues } from '@/lib/data/venues';
 import { parseAsJson, useQueryState } from 'nuqs';
+import type { FormEvent } from 'react';
 import { z } from 'zod';
 import { NumericFormat } from '@marigold/system';
+
+// Handling form data
+// ---------------
+export const getFormData = (e: FormEvent<HTMLFormElement>) => {
+  const data = new FormData(e.currentTarget);
+  const result: Record<string, FormDataEntryValue | FormDataEntryValue[]> = {};
+
+  for (const [key, value] of data.entries()) {
+    if (result[key]) {
+      result[key] = Array.isArray(result[key])
+        ? [...(result[key] as FormDataEntryValue[]), value]
+        : [result[key] as FormDataEntryValue, value];
+    } else {
+      result[key] = value;
+    }
+  }
+
+  return result;
+};
 
 // URL
 // ---------------
@@ -9,6 +29,7 @@ const urlSchema = z.object({
   type: z.number().optional(),
   capacity: z.number().optional(),
   price: z.number().optional(),
+  traits: z.array(z.string()).optional(),
   rating: z.number().optional(),
 });
 
@@ -18,6 +39,7 @@ export const defaultFilter = {
   type: undefined,
   capacity: Math.max(...venues.map(venue => venue.capacity)),
   price: Math.max(...venues.map(venue => venue.price.to)),
+  traits: venueTraits,
   rating: undefined,
 } satisfies VenueFilter;
 
@@ -28,6 +50,9 @@ const formSchema = z.object({
   type: z.string(),
   capacity: z.string(),
   price: z.string(),
+  traits: z
+    .union([z.string(), z.array(z.string())])
+    .transform(value => (Array.isArray(value) ? value : [value])),
   rating: z.string(),
 });
 
@@ -39,6 +64,7 @@ export const toFormSchema = urlSchema.transform(data => ({
   type: String(data.type ?? ''),
   capacity: data.capacity ?? defaultFilter.capacity,
   price: data.price ?? defaultFilter.price,
+  traits: data.traits ?? defaultFilter.traits,
   rating: String(data.rating ?? ''),
 })).parse;
 
@@ -47,6 +73,7 @@ export const toUrlSchema = formSchema.transform(data => ({
   type: data.type ? Number(data.type) : undefined,
   capacity: data.capacity ? Number(data.capacity) : undefined,
   price: data.price ? Number(data.price) : undefined,
+  traits: data.traits.length > 0 ? data.traits : undefined,
   rating: data.rating ? Number(data.rating) : undefined,
 })).safeParse;
 
@@ -55,12 +82,12 @@ export const toDisplayValue = {
   type: (value: number) => `Type: ${venueTypes[value] ?? 'Unknown'}`,
   capacity: (value: number) => (
     <>
-      Capacity: <NumericFormat value={value} />
+      Min Capacity: <NumericFormat value={value} />
     </>
   ),
   price: (value: number) => (
     <>
-      Price:{' '}
+      Max. Price:{' '}
       <NumericFormat
         style="currency"
         value={value}
@@ -69,19 +96,40 @@ export const toDisplayValue = {
       />
     </>
   ),
+  traits: (value: string[]) => (
+    <>
+      Traits:{' '}
+      {value.length <= 3
+        ? value.join(', ')
+        : `${value.slice(0, 2).join(', ')} (+${value.length - 2} more)`}
+    </>
+  ),
   rating: (value: number) => (
     <>
-      Rating: <NumericFormat value={value} minimumFractionDigits={1} />
+      Min. Rating: <NumericFormat value={value} minimumFractionDigits={1} />
     </>
   ),
 };
 
 // Hooks
 // ---------------
-export const useFilter = () =>
-  useQueryState(
+type FilterKeys = keyof typeof defaultFilter;
+
+export const useFilter = () => {
+  const [filter, setFilter] = useQueryState(
     'filter',
     parseAsJson(urlSchema.parse).withDefault(defaultFilter)
   );
+
+  const removeFilter = (keys: Set<FilterKeys>) => {
+    const next = { ...filter };
+    keys.forEach(key => {
+      next[key] = defaultFilter[key] as any;
+    });
+    setFilter(next);
+  };
+
+  return { filter, setFilter, removeFilter } as const;
+};
 
 export const useSearch = () => useQueryState('q', { defaultValue: '' });