+++ docs/app/(examples)/pattern/wrapper.tsx
@@ -3,6 +3,7 @@
 import { ruiTheme } from '@/theme';
 import { MarigoldProvider, OverlayContainerProvider } from '@/ui';
 import type { PropsWithChildren } from 'react';
+import { cn } from '@marigold/system';
 
 export interface WrapperProps {
   className?: string;
@@ -14,11 +15,15 @@ export const Wrapper = ({
 }: PropsWithChildren<WrapperProps>) => {
   return (
     <div data-theme="rui">
-      <OverlayContainerProvider value="portalContainer">
-        <MarigoldProvider className={className} theme={ruiTheme}>
-          <div className="h-[calc(100dvh-var(--page-header-height))] p-(--page-padding) md:p-(--page-padding-md) xl:p-(--page-padding-xl)">
-            {children}
-          </div>
+      <OverlayContainerProvider container="portalContainer">
+        <MarigoldProvider
+          className={cn(
+            'p-(--page-padding) md:p-(--page-padding-md) xl:p-(--page-padding-xl)',
+            className
+          )}
+          theme={ruiTheme}
+        >
+          {children}
         </MarigoldProvider>
       </OverlayContainerProvider>
     </div>