+++ themes/theme-rui/src/components/Button.styles.ts
@@ -4,7 +4,7 @@ import { cva } from '@marigold/system';
 export const Button: ThemeComponent<'Button'> = cva(
   [
     'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-[color,box-shadow]',
-    '[&_svg]:shrink-0',
+    '[&_svg]:pointer-events-none [&_svg]:shrink-0',
     'focus-visible:util-focus-ring outline-none disabled:util-disabled',
     'pending:text-disabled-foreground pending:bg-disabled pending:cursor-not-allowed pending:border-none',
     'cursor-pointer',
@@ -12,16 +12,12 @@ export const Button: ThemeComponent<'Button'> = cva(
   {
     variants: {
       variant: {
-        primary:
-          'bg-brand text-brand-foreground shadow-sm shadow-black/5 hover:bg-brand/90',
+        primary: 'bg-brand text-brand-foreground shadow-xs hover:bg-brand/90',
         secondary:
-          'border border-input bg-background shadow-sm shadow-black/5 hover:bg-hover hover:text-foreground expanded:bg-hover',
+          'border border-input bg-background shadow-xs hover:bg-hover hover:text-foreground expanded:bg-hover',
         destructive:
           'bg-destructive text-destructive-foreground hover:bg-destructive/90',
         ghost: 'hover:bg-hover hover:text-foreground',
-        // only used for calendar in MonthListBox and YearListBox to have a hover over other options
-        // TODO: Remove this in future and rethink the MonthListBox and YearListBox or use another variant
-        text: 'hover:bg-hover',
       },
       size: {
         default: 'h-button px-4 py-2 [&_svg]:size-4',