+++ themes/theme-docs/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @marigold/theme-docs
 
+## 3.1.3
+
+### Patch Changes
+
+- 81f1c9d: fix broken release
+- Updated dependencies [424e2f4]
+- Updated dependencies [81f1c9d]
+  - @marigold/components@14.1.1
+  - @marigold/system@14.1.1
+  - @marigold/theme-plugins@1.0.2
+
+## 3.1.2
+
+### Patch Changes
+
+- Updated dependencies [cc493fc]
+- Updated dependencies [930e633]
+- Updated dependencies [8f550ec]
+- Updated dependencies [69e7b61]
+- Updated dependencies [ea0f758]
+- Updated dependencies [8e178b7]
+- Updated dependencies [2163518]
+- Updated dependencies [37f40ba]
+  - @marigold/components@14.1.0
+  - @marigold/system@14.1.0
+
 ## 3.1.1
 
 ### Patch Changes
@@ -319,7 +345,6 @@
 ### Minor Changes
 
 - 6687af7: refa: remove footer from `<Dialog>` + allow styles
-
   - dialogs can only have action now
   - align buttons in `<Dialog.Actions>` correctly